Refactor to break out signal handler and utils. Create header files for completeness and remove some unecesary includes
