+++ Makefile
@@ -7,9 +7,9 @@ LDFLAGS = -lpthread
 TARGET = p2p_chat
 
 # Source files and object files
-SRCS = peer.c network.c protocol.c ui_terminal.c discovery.c
+SRCS = peer.c network.c protocol.c ui_terminal.c discovery.c utils.c signal_handler.c
 OBJS = $(SRCS:.c=.o)
-HEADERS = peer.h ui_terminal.h discovery.h
+HEADERS = peer.h ui_terminal.h discovery.h utils.h signal_handler.h
 
 # Default target
 all: $(TARGET)
