+++ peer.c
@@ -9,14 +9,6 @@
 // which holds information about the peer that is running)
 app_state_t *g_state = NULL;
 
-// Used to catch SIGINT and SIGTERM signals and perform graceful shutdown via g_state used in a loop
-void handle_signal(int sig) {
-    if (g_state) {
-        g_state->running = 0;
-    }
-    log_message("Received signal %d. Shutting down...", sig);
-}
-
 // Initialise application state
 void init_app_state(app_state_t *state, const char *username) {
     // Init the struct to hold state with zeros
@@ -65,29 +57,6 @@ void cleanup_app_state(app_state_t *state) {
     // use the mutex after it's destroyed.
     pthread_mutex_destroy(&state->peers_mutex);
     log_message("Cleanup complete");
-}
-
-/*
- * This function creates a timestamped log entry and outputs it to the console.
- * It works similar to printf() but automatically adds a timestamp prefix in the
- * format [HH:MM:SS] and a newline character at the end of each message.
- * The function also ensures immediate output by flushing stdout.
-*/
-void log_message(const char *format, ...) {
-    time_t now = time(NULL);
-    char time_str[20];
-    strftime(time_str, sizeof(time_str), "%H:%M:%S", localtime(&now));
-    
-    printf("[%s] ", time_str);
-    
-    va_list args;
-    va_start(args, format);
-    // print out the args using the format string, like printf
-    vprintf(format, args);
-    va_end(args);
-    
-    printf("\n");
-    fflush(stdout);
 }
 
 /*
