+++ utils.h
@@ -1,0 +1,10 @@
+// utils.h
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <time.h>
+
+// Function declarations
+void log_message(const char *format, ...);
+
+#endif // UTILS_H
